CrossOrigin 추가 및 Vue.js HTTP통신(Axios)를 위한 코드수정(세분화)

+++ Ch10/src/main/java/kr/ch10/service/User1Service.java
@@ -14,19 +14,46 @@ public class User1Service {
 	@Autowired
 	private User1Repository repo;
 	
-	public void insertUser1(User1Entity user1) {
-		repo.save(user1);
+	public int insertUser1(User1Entity user1) {
+		// repo.save(user1);
+		// 해당 user id가 존재할 경우를 위한 세분화
+		if(!repo.existsById(user1.getId())) {
+			repo.save(user1);
+			return 0;
+		} else {
+			return 1;
+		}
 	}
 	public User1Entity selectUser1(String id) {
-		return repo.findById(id).get();
+		// return repo.findById(id).get();
+		// 해당 user id가 존재하지 않을 경우를 위한 세분화
+		if(repo.existsById(id))
+			return repo.findById(id).get();
+		else
+			return null;
 	}
 	public List<User1Entity> selectUser1s() {
 		return repo.findAll();
 	}
-	public void updateUser1(User1Entity user1) {
-		repo.save(user1);
-	}
-	public void deleteUser1(String id) {
-		repo.deleteById(id);
-	}
+	public int updateUser1(User1Entity user1) {
+		// repo.save(user1);
+		// 해당 user id가 존재하지 않을 경우를 위한 세분화
+		if(repo.existsById(user1.getId())) {
+			repo.save(user1);
+			return 0;
+		} else {
+			return 1;
+		}
+    }
+	public int deleteUser1(String id) {
+		// repo.deleteById(id);
+		// 해당 user id가 존재하지 않을 경우를 위한 세분화
+		if(repo.existsById(id)) {
+			repo.deleteById(id);
+			return 0;
+		}
+		else {
+			return 1;
+		}
+    }
 }
